requirements.txt
Flask==2.0.2
Flask-Mail==0.9.1
flask-wtf==1.0.1
mysql-connector-python==8.0.26
pymysql==1.0.2
bcrypt==3.2.0
APScheduler==3.9.1
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.23.0
python-dateutil==2.8.2
requests==2.26.0
Flask==2.0.2
werkzeug==2.0.3
dash==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-bootstrap-components==1.0.0
